+++ CHANGELOG.rst
@@ -14,6 +14,13 @@ Change Log
 Unreleased
 ~~~~~~~~~~
 
+[3.5.0] - 2021-02-18
+~~~~~~~~~~~~~~~~~~~~
+* Add new UI for instructor dashboard that groups attempts for each user and exam.
+* Add endpoint that returns a list of most recent attempts for each user and exam. Each
+  attempt that is returned contains additional data on the past attempts
+  associated with the user/exam.
+
 [3.4.1] - 2021-02-17
 ~~~~~~~~~~~~~~~~~~~~
 * Restrict the resume option on the instructor dashboard to attempts that are
@@ -22,7 +29,7 @@ Unreleased
 [3.4.0] - 2021-02-11
 ~~~~~~~~~~~~~~~~~~~~
 * Add a new interstitial for exam attempts in the "ready_to_resume" state to
-  indicate to learner that their exam attempt is ready to be resumed and to 
+  indicate to learner that their exam attempt is ready to be resumed and to
   prompt the learner to resume their exam.
 
 [3.3.0] - 2021-02-11